Feature - Users, Login & Register

+++ domain/src/main/java/com/example/domain/usecase/DeleteProductUseCase.kt
@@ -1,8 +1,7 @@
 package com.example.domain.usecase
 
-import com.example.domain.model.CartItemModel
 import com.example.domain.repository.CartRepository
 
 class DeleteProductUseCase(private val cartRepository: CartRepository) {
-    suspend fun execute(cartItemId: Int, userId: Int) = cartRepository.deleteItem(cartItemId, userId)
+    suspend fun execute(cartItemId: Int, userId: Long) = cartRepository.deleteItem(cartItemId, userId)
 }
